# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Djam's Food & Drink Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Worker {
  id: number;
  name: string;
  dailyRate: number;
  role: string;
  createdAt: string;
}

export interface CreateWorkerBody {
  name: string;
  dailyRate: number;
  role: string;
}

export interface WorkerBalance {
  workerId: number;
  workerName: string;
  totalWorkUnits: number;
  dailyRate: number;
  totalEarned: number;
  totalAdvances: number;
  totalPaidOut: number;
  currentBalance: number;
}

export interface Attendance {
  id: number;
  workerId: number;
  date: string;
  /** 1.0 = full day, 0.5 = half day, 0.25 = 2 hours, etc. */
  units: number;
  note?: string;
  createdAt: string;
}

export interface CreateAttendanceBody {
  workerId: number;
  date: string;
  units: number;
  note?: string;
}

export interface Advance {
  id: number;
  workerId: number;
  amount: number;
  note?: string;
  createdAt: string;
}

export interface CreateAdvanceBody {
  workerId: number;
  amount: number;
  note?: string;
}

export interface Payout {
  id: number;
  workerId: number;
  amount: number;
  note?: string;
  createdAt: string;
}

export interface CreatePayoutBody {
  workerId: number;
  amount: number;
  note?: string;
}

export interface Supplier {
  id: number;
  supplierName: string;
  item: string;
  amount: number;
  date: string;
  note?: string;
  createdAt: string;
}

export interface CreateSupplierBody {
  supplierName: string;
  item: string;
  amount: number;
  date: string;
  note?: string;
}

export interface MonthlyTotal {
  month: string;
  total: number;
}

export interface RevenueEntry {
  id: number;
  date: string;
  amount: number;
  note?: string;
  createdAt: string;
}

export interface CreateRevenueBody {
  date: string;
  amount: number;
  note?: string;
}

export type MenuItemCategory =
  (typeof MenuItemCategory)[keyof typeof MenuItemCategory];

export const MenuItemCategory = {
  pizza: "pizza",
  burger: "burger",
  drink: "drink",
  sandwich: "sandwich",
  dessert: "dessert",
  other: "other",
} as const;

export interface MenuItem {
  id: number;
  name: string;
  category: MenuItemCategory;
  price: number;
  ingredients: string;
  available: boolean;
  createdAt: string;
}

export type CreateMenuItemBodyCategory =
  (typeof CreateMenuItemBodyCategory)[keyof typeof CreateMenuItemBodyCategory];

export const CreateMenuItemBodyCategory = {
  pizza: "pizza",
  burger: "burger",
  drink: "drink",
  sandwich: "sandwich",
  dessert: "dessert",
  other: "other",
} as const;

export interface CreateMenuItemBody {
  name: string;
  category: CreateMenuItemBodyCategory;
  price: number;
  ingredients: string;
  available?: boolean;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  ready: "ready",
  completed: "completed",
} as const;

export interface OrderItem {
  menuItemId: number;
  name: string;
  quantity: number;
  price: number;
}

export interface Order {
  id: number;
  tableNumber: string;
  items: OrderItem[];
  status: OrderStatus;
  note?: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateOrderBodyItemsItem = {
  menuItemId: number;
  quantity: number;
};

export interface CreateOrderBody {
  tableNumber: string;
  items: CreateOrderBodyItemsItem[];
  note?: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  ready: "ready",
  completed: "completed",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface DashboardSummary {
  month: string;
  totalRevenue: number;
  totalSupplierCosts: number;
  totalAdvances: number;
  totalPayouts: number;
  netProfit: number;
  totalWorkersCount: number;
  pendingOrdersCount: number;
}

export type ListAttendanceParams = {
  workerId?: number;
  /**
   * YYYY-MM format
   */
  month?: string;
};

export type ListAdvancesParams = {
  workerId?: number;
  month?: string;
};

export type ListPayoutsParams = {
  workerId?: number;
};

export type ListSuppliersParams = {
  /**
   * YYYY-MM format
   */
  month?: string;
};

export type GetSuppliersMonthlyTotalParams = {
  month: string;
};

export type ListRevenueParams = {
  month?: string;
};

export type ListMenuItemsParams = {
  category?: string;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  ready: "ready",
  completed: "completed",
} as const;

export type GetDashboardSummaryParams = {
  /**
   * YYYY-MM
   */
  month: string;
};